Revert "Bug 1966196 - Expand EnvironmentCoordinate hops bitwidth to 16 r=nbp" for causing spidermonkey bustages.
This reverts commit f0c591856b1421d38a70231f3912c494309689e4.

+++ js/src/frontend/EmitterScope.h
@@ -57,7 +57,7 @@ class MOZ_STACK_CLASS EmitterScope : public Nestable<EmitterScope> {
 #endif
 
   // The number of enclosing environments. Used for error checking.
-  uint16_t environmentChainLength_;
+  uint8_t environmentChainLength_;
 
   // The next usable slot on the frame for not-closed over bindings.
   //
